# docs/swagger.yaml
basePath: /api
definitions:
  add_member.AddMemberToGroupRequest:
    properties:
      inviteeUserId:
        type: string
    required:
    - inviteeUserId
    type: object
...
  get_job_for_user.JobResponse:
    properties:
      id:
        type: string
      is_assigned:
        type: boolean
      name:
        type: string
      permissions:
        type: integer
    type: object
...
  list_messages.GroupChatMessageResponse:
    properties:
      author:
        $ref: '#/definitions/list_messages.GroupMessageAuthorResponse'
      content:
        type: string
      createdAt:
        type: string
      groupId:
        type: string
      id:
        type: string
      reactions:
        items:
          $ref: '#/definitions/list_messages.GroupMessageReactionResponse'
        type: array
    type: object
  list_messages.GroupMessageAuthorResponse:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      userId:
        type: string
    type: object
  list_messages.GroupMessageReactionResponse:
    properties:
      id:
        type: string
      reaction:
        type: string
      users:
        items:
          $ref: '#/definitions/list_messages.GroupMessageReactionUserResponse'
        type: array
    type: object
  list_messages.GroupMessageReactionUserResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  models.KickMemberRequest:
    properties:
      member_id:
        example: user-123
        type: string
    type: object
...
  models.UnassignJobRequest:
    properties:
      job_id:
        example: job-456
        type: string
      user_id:
        example: user-123
        type: string
    type: object
  models.UnassignJobResponse:
    properties:
      message:
        example: Rôle retiré avec succès
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  vote_option_poll.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@supchat.io
    name: Support API SupChat
  description: API pour le service de messagerie SupChat
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SupChat API
  version: "1.0"
paths:
...
  /api/groups/{group_id}/members:
    delete:
      consumes:
      - application/json
      description: Expulse un utilisateur spécifique d'un groupe de discussion
      parameters:
      - description: ID du groupe
        in: path
        name: group_id
        required: true
        type: string
      - description: ID du membre à expulser
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.KickMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Membre expulsé avec succès
        "400":
          description: Erreur de paramètre
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Non autorisé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Expulser un membre du groupe
      tags:
      - group
    post:
      consumes:
      - application/json
      description: Ajoute un utilisateur spécifique à un groupe de discussion
      parameters:
      - description: ID du groupe
        in: path
        name: group_id
        required: true
        type: string
      - description: Informations de l'utilisateur à ajouter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/add_member.AddMemberToGroupRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Membre ajouté avec succès
        "400":
          description: Erreur de paramètre ou impossibilité de s'ajouter soi-même
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Non autorisé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Ajouter un membre au groupe
      tags:
      - group
  /api/groups/{group_id}/messages:
    get:
      consumes:
      - application/json
      description: Récupère les messages d'un groupe spécifique
      parameters:
      - description: ID du groupe
        in: path
        name: group_id
        required: true
        type: string
      - default: 20
        description: Nombre maximum de messages à récupérer
        in: query
        maximum: 100
        name: limit
        type: integer
      - description: Récupérer les messages avant cette date (format ISO)
        in: query
        name: before
        type: string
      - description: Récupérer les messages après cette date (format ISO)
        in: query
        name: after
        type: string
      - description: Récupérer les messages autour de ce message
        in: query
        name: aroundMessageId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Liste des messages
          schema:
            items:
              $ref: '#/definitions/list_messages.GroupChatMessageResponse'
            type: array
        "400":
          description: Requête invalide
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Non autorisé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Lister les messages d'un groupe
      tags:
      - group-chat
...
  /api/job/unassign:
    post:
      consumes:
      - application/json
      description: Retire un rôle de travail spécifié d'un utilisateur
      parameters:
      - description: Informations pour retirer un rôle
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UnassignJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Opération réussie
          schema:
            $ref: '#/definitions/models.UnassignJobResponse'
        "400":
          description: Erreur de paramètre
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Non autorisé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Désassigner un rôle de travail
      tags:
      - job
  /api/job/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Récupère la liste des rôles de travail avec leurs permissions pour
        un utilisateur spécifique
      parameters:
      - description: ID de l'utilisateur
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Liste des rôles de travail de l'utilisateur
          schema:
            $ref: '#/definitions/get_job_for_user.JobResponse'
        "400":
          description: Erreur de paramètre
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Non autorisé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtenir les rôles de travail d'un utilisateur
      tags:
      - job
...
  /api/workspaces/{workspace_id}/poll/{poll_id}/vote/{option_id}:
    post:
      consumes:
      - application/json
      description: Enregistre le vote d'un utilisateur pour une option spécifique
        d'un sondage
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      - description: ID du sondage
        in: path
        name: poll_id
        required: true
        type: string
      - description: ID de l'option
        in: path
        name: option_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Vote enregistré avec succès
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Erreur de paramètre
          schema:
            $ref: '#/definitions/vote_option_poll.ErrorResponse'
        "401":
          description: Non autorisé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/vote_option_poll.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Voter pour une option
      tags:
      - poll
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"